lvm2: stop depending on util-linux
util-linux was made an optional dependency to avoid infinite recursion
on pkgsStatic back in https://github.com/NixOS/nixpkgs/pull/169343 but
since then infinite recursion was reintroduced in the non-static pkgs
too due to the fact that we added cryptsetup support to systemd

To have this util-linux dependency just to fix up some paths in
blkdeactivate which is a script with no real use in NixOS as far as I
know seems overkill to me.

Lets just remove the dependency in general.

It makes the dependency chain for systemd less complicated.